ad9656:Add driver for ad9656 chip
The AD9656 chip contains 4 16 bit ADCs with a sampling rate up to 125MSPS.
Data from the ADCs is sent to the carrier board through 4 JESD204 lines.
The chip contains a series of control registers that can be accessed
through the SPI interface. The chip can generate test patterns for
determining if the JESD204 connection is functioning correctly and also
user test patterns. The user test patterns are configured in the
configuration registers and are sent on the JESD204 link instead of the
captured ADC data.

Signed-off-by: Dan Hotoleanu <dan.hotoleanu@analog.com>